+++ lib/extended/extended.dart
@@ -2,11 +2,10 @@ import 'dart:ui';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_waya/flutter_waya.dart';
 
-export 'extended_widgets.dart';
-export 'global_options.dart';
 export 'icon.dart';
 export 'list_wheel.dart';
 export 'overlay/overlay.dart';
@@ -16,6 +15,7 @@ export 'tab_bar.dart';
 export 'text.dart';
 export 'text_field.dart';
 export 'universal.dart';
+export 'widgets_app.dart';
 
 class SystemUiOverlayStyleLight extends SystemUiOverlayStyle {
   const SystemUiOverlayStyleLight(
@@ -37,156 +37,171 @@ class SystemUiOverlayStyleDark extends SystemUiOverlayStyle {
       super.statusBarBrightness = Brightness.light});
 }
 
-extension ExtensionWidgetMethod on Widget {
-  Future<T?> push<T extends Object?, TO extends Object?>(
-      {bool maintainState = true,
-      bool fullscreenDialog = false,
-      RoutePushStyle pushStyle = RoutePushStyle.material,
-      RouteSettings? settings,
-      bool replacement = false,
-      TO? result}) {
-    if (replacement) {
-      return pushReplacement(
-          settings: settings,
-          maintainState: maintainState,
-          fullscreenDialog: fullscreenDialog,
-          pushStyle: pushStyle,
-          result: result);
-    } else {
-      return GlobalOptions().globalNavigatorKey.currentState!.push(
-          buildPageRoute(
-              maintainState: maintainState,
-              fullscreenDialog: fullscreenDialog,
-              settings: settings,
-              pushStyle: pushStyle));
-    }
+class GlobalOptions {
+  factory GlobalOptions() => _singleton ??= GlobalOptions._();
+
+  GlobalOptions._();
+
+  static GlobalOptions? _singleton;
+
+  WidgetsBinding widgetsBinding = WidgetsBinding.instance;
+
+  SchedulerBinding schedulerBinding = SchedulerBinding.instance;
+
+  /// 统一全局控制 是否可返回
+  /// true 允许的返回
+  /// false 不允许返回
+  ///  [ExtendedScaffold.onWillPop] 方法优先于 [scaffoldWillPop]
+  bool _scaffoldWillPop = true;
+
+  bool get scaffoldWillPop => _scaffoldWillPop;
+
+  void setScaffoldWillPop(bool value) => _scaffoldWillPop = value;
+
+  GlobalKey<NavigatorState> _globalNavigatorKey = GlobalKey<NavigatorState>();
+
+  GlobalKey<NavigatorState> get globalNavigatorKey => _globalNavigatorKey;
+
+  /// 设置全局 [NavigatorKey]
+  /// Set the global [NavigatorKey]
+  void setGlobalNavigatorKey(GlobalKey<NavigatorState> navigatorKey) {
+    _globalNavigatorKey = navigatorKey;
   }
 
-  /// 打开新页面替换当前页面
-  Future<T?> pushReplacement<T extends Object?, TO extends Object?>(
-          {bool maintainState = true,
-          bool fullscreenDialog = false,
-          RoutePushStyle pushStyle = RoutePushStyle.material,
-          RouteSettings? settings,
-          TO? result}) =>
-      GlobalOptions().globalNavigatorKey.currentState!.pushReplacement(
-          buildPageRoute(
-              settings: settings,
-              maintainState: maintainState,
-              fullscreenDialog: fullscreenDialog,
-              pushStyle: pushStyle),
-          result: result);
+  GlobalKey<ScaffoldMessengerState>? _globalScaffoldMessengerKey;
 
-  /// 打开新页面 并移出堆栈所有页面
-  Future<T?> pushAndRemoveUntil<T extends Object?>(
-          {bool maintainState = true,
-          bool fullscreenDialog = false,
-          RoutePushStyle pushStyle = RoutePushStyle.material,
-          RouteSettings? settings,
-          RoutePredicate? predicate}) =>
-      GlobalOptions().globalNavigatorKey.currentState!.pushAndRemoveUntil(
-          buildPageRoute(
-              settings: settings,
-              maintainState: maintainState,
-              fullscreenDialog: fullscreenDialog,
-              pushStyle: pushStyle),
-          predicate ?? (_) => false);
+  GlobalKey<ScaffoldMessengerState>? get globalScaffoldMessengerKey =>
+      _globalScaffoldMessengerKey;
 
-  /// [ExtendedOverlay().showOverlay()]
-  ExtendedOverlayEntry? showOverlay({bool autoOff = false}) =>
-      ExtendedOverlay().showOverlay(this, autoOff: autoOff);
-
-  /// [ExtendedOverlay().showLoading(this)]
-  ExtendedOverlayEntry? showLoading({ModalWindowsOptions? options}) =>
-      Loading(custom: this, options: options).show();
-
-  /// [showDialogPopup]
-  Future<T?> showDialogPopup<T>({
-    /// 这个参数是一个方法,入参是 context,animation,secondaryAnimation,返回一个 Widget
-    RoutePageBuilder? builder,
-
-    /// GeneralDialog 配置
-    GeneralDialogOptions? options,
-  }) {
-    options ??= GlobalOptions().dialogOptions;
-    RouteTransitionsBuilder? transitionBuilder;
-    if (options.fromStyle != PopupFromStyle.fromCenter) {
-      transitionBuilder = options.transitionBuilder ??
-          (__, Animation<double> animation, _, Widget child) {
-            late Offset translation;
-            switch (options!.fromStyle) {
-              case PopupFromStyle.fromLeft:
-                translation = Offset(animation.value - 1, 0);
-                break;
-              case PopupFromStyle.fromRight:
-                translation = Offset(1 - animation.value, 0);
-                break;
-              case PopupFromStyle.fromTop:
-                translation = Offset(0, animation.value - 1);
-                break;
-              case PopupFromStyle.fromBottom:
-                translation = Offset(0, 1 - animation.value);
-                break;
-              case PopupFromStyle.fromCenter:
-                translation = const Offset(0, 0);
-                break;
-            }
-            return FractionalTranslation(
-                translation: translation, child: child);
-          };
-    }
-    return showGeneralDialog(
-        context: GlobalOptions().globalNavigatorKey.currentContext!,
-        pageBuilder: builder ?? (_, Animation<double> animation, __) => this,
-        barrierDismissible: options.barrierDismissible,
-        barrierLabel: options.barrierLabel,
-        barrierColor: options.barrierColor,
-        transitionDuration: options.transitionDuration,
-        transitionBuilder: transitionBuilder,
-        useRootNavigator: options.useRootNavigator,
-        routeSettings: options.routeSettings);
+  /// 设置全局 [ScaffoldMessengerKey]
+  /// Set the global [ScaffoldMessengerKey]
+  void setGlobalScaffoldMessengerKey(
+      GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey) {
+    _globalScaffoldMessengerKey = scaffoldMessengerKey;
   }
 
-  /// showModalBottomSheet
-  /// 关闭 closePopup()
-  Future<T?> showBottomPopup<T>(
-      {WidgetBuilder? builder, BottomSheetOptions? options}) {
-    options ??= GlobalOptions().bottomSheetOptions;
-    return showModalBottomSheet(
-        context: GlobalOptions().globalNavigatorKey.currentContext!,
-        builder: builder ?? toWidgetBuilder,
-        backgroundColor: options.backgroundColor,
-        elevation: options.elevation,
-        shape: options.shape,
-        clipBehavior: options.clipBehavior,
-        barrierColor: options.barrierColor,
-        routeSettings: options.routeSettings,
-        transitionAnimationController: options.transitionAnimationController,
-        isScrollControlled: options.isScrollControlled,
-        useRootNavigator: options.useRootNavigator,
-        isDismissible: options.isDismissible,
-        enableDrag: options.enableDrag);
+  RoutePushStyle _pushStyle = RoutePushStyle.cupertino;
+
+  RoutePushStyle get pushStyle => _pushStyle;
+
+  /// 设置全局路由跳转样式
+  /// Set the global route push style
+  void setGlobalPushMode(RoutePushStyle style) {
+    _pushStyle = style;
   }
 
-  /// showCupertinoModalPopup
-  /// 关闭 closePopup()
-  /// 全屏显示
-  Future<T?> showCupertinoBottomPopup<T>({
-    WidgetBuilder? builder,
-    bool useRootNavigator = true,
-    ImageFilter? filter,
-    Color barrierColor = kCupertinoModalBarrierColor,
-    bool barrierDismissible = true,
-    bool? semanticsDismissible,
-    RouteSettings? routeSettings,
-  }) =>
-      showCupertinoModalPopup(
-          context: GlobalOptions().globalNavigatorKey.currentContext!,
-          builder: builder ?? toWidgetBuilder,
-          filter: filter,
-          barrierColor: barrierColor,
-          barrierDismissible: barrierDismissible,
-          semanticsDismissible: semanticsDismissible,
-          routeSettings: routeSettings,
-          useRootNavigator: useRootNavigator);
+  final EventBus _eventBus = EventBus();
+
+  EventBus get eventBus => _eventBus;
+
+  ModalWindowsOptions _modalWindowsOptions = const ModalWindowsOptions();
+
+  ModalWindowsOptions get modalWindowsOptions => _modalWindowsOptions;
+
+  /// 设置全局 ModalWindowsOptions 配置
+  /// Set the global [ModalWindowsOptions] 配置
+  void setModalWindowsOptions(ModalWindowsOptions modalWindowsOptions) {
+    _modalWindowsOptions = modalWindowsOptions;
+  }
+
+  LoadingOptions _loadingOptions = const LoadingOptions();
+
+  LoadingOptions get loadingOptions => _loadingOptions;
+
+  /// 设置全局 [LoadingOptions] 配置
+  /// Set the global [LoadingOptions] configuration
+  void setLoadingOptions(LoadingOptions options) {
+    _loadingOptions = options;
+  }
+
+  ToastOptions _toastOptions = const ToastOptions();
+
+  ToastOptions get toastOptions => _toastOptions;
+
+  /// 设置全局 [ToastOptions] 配置
+  /// Set the global [ToastOptions] configuration
+  void setToastOptions(ToastOptions options) {
+    _toastOptions = options;
+  }
+
+  BottomSheetOptions _bottomSheetOptions = const BottomSheetOptions();
+
+  BottomSheetOptions get bottomSheetOptions => _bottomSheetOptions;
+
+  /// 设置全局 [BottomSheet] 配置
+  /// Set the global [BottomSheet] configuration
+  void setBottomSheetOptions(BottomSheetOptions options) {
+    _bottomSheetOptions = options;
+  }
+
+  GeneralDialogOptions _dialogOptions = const GeneralDialogOptions();
+
+  GeneralDialogOptions get dialogOptions => _dialogOptions;
+
+  /// 设置全局 [GeneralDialogOptions] 配置
+  /// Set the global [GeneralDialogOptions] configuration
+  void setGeneralDialogOptions(GeneralDialogOptions options) {
+    _dialogOptions = options;
+  }
+
+  WheelOptions _wheelOptions = const WheelOptions();
+
+  WheelOptions get wheelOptions => _wheelOptions;
+
+  /// 设置全局 [ListWheel] 配置
+  /// Set the global [ListWheel] configuration
+  void setWheelOptions(WheelOptions options) {
+    _wheelOptions = options;
+  }
+
+  PickerWheelOptions _pickerWheelOptions = const PickerWheelOptions();
+
+  PickerWheelOptions get pickerWheelOptions => _pickerWheelOptions;
+
+  /// 设置全局 [PickerWheelOptions] 配置
+  /// Set the global [PickerWheelOptions] configuration
+  void setPickerWheelOptions(PickerWheelOptions options) {
+    _pickerWheelOptions = options;
+  }
+
+  bool _logHasDottedLine = true;
+
+  bool get logHasDottedLine => _logHasDottedLine;
+
+  void setLogDottedLine(bool has) => _logHasDottedLine = has;
+
+  Header globalRefreshHeader = const ClassicHeader(
+      dragText: '请尽情拉我',
+      armedText: '可以松开我了',
+      readyText: '我要开始刷新了',
+      processingText: '我在拼命刷新中',
+      processedText: '我已经刷新完成了',
+      failedText: '我刷新失败了唉',
+      noMoreText: '没有更多了',
+      showMessage: false);
+
+  Footer globalRefreshFooter = const ClassicFooter(
+      dragText: '请尽情拉我',
+      armedText: '可以松开我了',
+      readyText: '我要准备加载了',
+      processingText: '我在拼命加载中',
+      processedText: '我已经加载完成了',
+      failedText: '我加载失败了唉',
+      noMoreText: '没有更多了哦',
+      showMessage: false);
 }
+
+void addPostFrameCallback(FrameCallback duration) =>
+    GlobalOptions().widgetsBinding.addPostFrameCallback(duration);
+
+void addObserver(WidgetsBindingObserver observer) =>
+    GlobalOptions().widgetsBinding.addObserver(observer);
+
+void removeObserver(WidgetsBindingObserver observer) =>
+    GlobalOptions().widgetsBinding.removeObserver(observer);
+
+void addPersistentFrameCallback(FrameCallback duration) =>
+    GlobalOptions().widgetsBinding.addPersistentFrameCallback(duration);
+
+void addTimingsCallback(TimingsCallback callback) =>
+    GlobalOptions().widgetsBinding.addTimingsCallback(callback);